Cleanup / Fix warnings

+++ crostouchpad/smbus.c
@@ -11,7 +11,7 @@ static void outb_p(unsigned char value, unsigned short int port) {
 	__outbyte(port, value);
 }
 
-static void cyapa_set_slave_addr(PCYAPA_CONTEXT pDevice, int smbusread) {
+static void cyapa_set_slave_addr(PCYAPA_CONTEXT pDevice, uint8_t smbusread) {
 	outb_p(0xce + smbusread, SMBHSTADD(pDevice)); // 0x67
 }
